add navlinks
