+++ package.json
@@ -8,8 +8,8 @@
     "@testing-library/user-event": "^12.8.1",
     "node-sass": "4.14.1",
     "react": "^17.0.1",
-    "react-router-dom": "^5.2.0",
     "react-dom": "^17.0.1",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
     "web-vitals": "^1.1.0"
   },
